Release/1.2

See merge request mobile/components/downloader!5

+++ downloader/src/main/java/com/apkpure/components/downloader/service/services/DownloadServiceAssistUtils.kt
@@ -10,11 +10,12 @@ import com.apkpure.components.downloader.db.enums.DownloadTaskStatusType
 import com.apkpure.components.downloader.service.DownloadManager
 import com.apkpure.components.downloader.service.misc.CustomDownloadListener4WithSpeed
 import com.apkpure.components.downloader.service.misc.DownloadTaskChangeLister
-import com.apkpure.components.downloader.service.misc.DownloadTaskDeleteLister
+import com.apkpure.components.downloader.service.misc.DownloadTaskFileChangeLister
 import com.apkpure.components.downloader.service.misc.TaskManager
 import com.apkpure.components.downloader.utils.*
 import com.liulishuo.okdownload.DownloadTask
 import io.reactivex.disposables.Disposable
+import java.io.File
 
 /**
  * @author xiongke
@@ -33,6 +34,7 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
     companion object {
         private const val EXTRA_PARAM_ACTION = "download_param_action"
         private const val EXTRA_PARAM_IS_DELETE = "is_delete"
+        private const val EXTRA_PARAM_FILE_NAMNE = "file_name"
         val downloadTaskLists = mutableListOf<DownloadTaskBean>()
 
         object ActionType {
@@ -43,6 +45,7 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
             const val ACTION_START_ALL = "start_all"
             const val ACTION_STOP_ALL = "stop_all"
             const val ACTION_DELETE_ALL = "delete_all"
+            const val ACTION_FILE_RENAME = "file_rename"
         }
 
         fun newInitIntent(mContext: Context, clazz: Class<*>): Intent {
@@ -78,6 +81,14 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                 this.action = ActionType.ACTION_DELETE_ALL
             }
         }
+
+        fun newRenameIntent(mContext: Context, clazz: Class<*>, taskUrl: String, fileName: String): Intent {
+            return Intent(mContext, clazz).apply {
+                this.action = ActionType.ACTION_FILE_RENAME
+                this.putExtra(EXTRA_PARAM_ACTION, taskUrl)
+                this.putExtra(EXTRA_PARAM_FILE_NAMNE, fileName)
+            }
+        }
     }
 
     private fun initial() {
@@ -97,7 +108,7 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                     this.totalLength = 0
                 }
                 updateDbDataAndNotify(this)
-                AppLogger.d(logTag, "onStart ${this.shortName} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
+                AppLogger.d(logTag, "onStart ${this.notificationTitle} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
             }
         }
 
@@ -107,7 +118,7 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                 this.totalLength = totalLength
                 this.absolutePath = task.file?.path
                 updateDbDataAndNotify(this)
-                AppLogger.d(logTag, "onInfoReady ${this.shortName} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
+                AppLogger.d(logTag, "onInfoReady ${this.notificationTitle} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
             }
         }
 
@@ -119,7 +130,7 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                 this.downloadTaskStatusType = downloadTaskStatusType
                 val downloadPercent = CommonUtils.formatPercentInfo(this.currentOffset, this.totalLength)
                 updateDbDataAndNotify(this)
-                AppLogger.d(logTag, "onProgress ${this.shortName} ${task.connectionCount} ${this.currentOffset} ${this.totalLength} $downloadPercent")
+                AppLogger.d(logTag, "onProgress ${this.notificationTitle} ${task.connectionCount} ${this.currentOffset} ${this.totalLength} $downloadPercent")
             }
         }
 
@@ -128,7 +139,7 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                 this.downloadTaskStatusType = downloadTaskStatusType
                 this.absolutePath = task.file?.path
                 updateDbDataAndNotify(this)
-                AppLogger.d(logTag, "onCancel ${this.shortName} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
+                AppLogger.d(logTag, "onCancel ${this.notificationTitle} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
             }
         }
 
@@ -137,7 +148,7 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                 this.downloadTaskStatusType = downloadTaskStatusType
                 this.absolutePath = task.file?.path
                 updateDbDataAndNotify(this)
-                AppLogger.d(logTag, "onSuccess ${this.shortName} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
+                AppLogger.d(logTag, "onSuccess ${this.notificationTitle} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
             }
         }
 
@@ -146,13 +157,13 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                 this.downloadTaskStatusType = downloadTaskStatusType
                 this.absolutePath = task.file?.path
                 updateDbDataAndNotify(this)
-                AppLogger.d(logTag, "onError ${this.shortName} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
+                AppLogger.d(logTag, "onError ${this.notificationTitle} ${task.connectionCount} ${this.currentOffset} ${this.totalLength}")
             }
         }
 
         override fun onRetry(downloadTaskBean: DownloadTaskBean?, task: DownloadTask, downloadTaskStatusType: DownloadTaskStatusType, retryCount: Int) {
             downloadTaskBean?.apply {
-                AppLogger.d(logTag, "onRetry ${this.shortName}  $retryCount")
+                AppLogger.d(logTag, "onRetry ${this.notificationTitle}  $retryCount")
             }
         }
     }
@@ -169,23 +180,28 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
             }
             ActionType.ACTION_STOP -> {
                 intent.getStringExtra(EXTRA_PARAM_ACTION)?.apply {
-                    stop(this, true)
+                    stop(this)
                 }
             }
             ActionType.ACTION_DELETE -> {
                 val isDeleteFile = intent.getBooleanExtra(EXTRA_PARAM_IS_DELETE, false)
                 intent.getStringExtra(EXTRA_PARAM_ACTION)?.apply {
-                    delete(this, isDeleteFile, true)
+                    delete(this, isDeleteFile)
                 }
             }
             ActionType.ACTION_START_ALL -> {
                 startAll()
             }
             ActionType.ACTION_STOP_ALL -> {
-                stopAll(true)
+                stopAll()
             }
             ActionType.ACTION_DELETE_ALL -> {
-                deleteAll(isDeleteFile = true, isCancelNotify = true)
+                deleteAll()
+            }
+            ActionType.ACTION_FILE_RENAME -> {
+                val taskUrl = intent.getStringExtra(EXTRA_PARAM_ACTION) ?: return
+                val fileName = intent.getStringExtra(EXTRA_PARAM_FILE_NAMNE) ?: return
+                renameTaskFile(taskUrl, fileName)
             }
         }
     }
@@ -215,32 +231,32 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
         DownloadManager.instance.getDownloadTask(downloadTaskBean.url) ?: let {
             downloadTaskLists.add(0, downloadTaskBean)
         }
-        TaskManager.instance.start(downloadTaskBean.url, downloadTaskBean.absolutePath)
+        TaskManager.instance.start(downloadTaskBean)
     }
 
     private fun startAll() {
         TaskManager.instance.startOnParallel()
     }
 
-    private fun stop(taskUrl: String, isCancelNotify: Boolean) {
+    private fun stop(taskUrl: String) {
         TaskManager.instance.stop(taskUrl)
-        if (isCancelNotify) {
-            DownloadManager.instance.getDownloadTask(taskUrl)?.let {
+        DownloadManager.instance.getDownloadTask(taskUrl)?.let {
+            if (it.showNotification) {
                 notifyHelper.notificationManager.cancel(it.notificationId)
             }
         }
     }
 
-    private fun stopAll(isCancelNotify: Boolean) {
+    private fun stopAll() {
         TaskManager.instance.stopAll()
         downloadTaskLists.forEach {
-            if (isCancelNotify) {
+            if (it.showNotification) {
                 notifyHelper.notificationManager.cancel(it.notificationId)
             }
         }
     }
 
-    private fun deleteAll(isDeleteFile: Boolean, isCancelNotify: Boolean) {
+    private fun deleteAll() {
         TaskManager.instance.deleteAll()
         AppDbHelper.instance.deleteAllTasks()
                 .compose(RxObservableTransformer.io_main())
@@ -249,12 +265,10 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                     override fun rxOnNext(t: Long) {
                         val missionList = downloadTaskLists
                         missionList.forEach {
-                            if (isCancelNotify) {
+                            if (it.showNotification) {
                                 notifyHelper.notificationManager.cancel(it.notificationId)
                             }
-                            if (isDeleteFile) {
-                                FsUtils.deleteFileOrDir(it.absolutePath)
-                            }
+                            FsUtils.deleteFileOrDir(it.absolutePath)
                         }
                         missionList.forEach {
                             DownloadTaskChangeLister.sendChangeBroadcast(mContext1, it.apply {
@@ -262,15 +276,21 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                             })
                         }
                         missionList.clear()
-                        DownloadTaskDeleteLister.sendAllDeleteBroadcast(mContext1)
+                        DownloadTaskFileChangeLister.sendAllDeleteBroadcast(mContext1, true)
                     }
 
-                    override fun rxOnError(e: Exception) = Unit
+                    override fun rxOnError(e: Exception) {
+                        DownloadTaskFileChangeLister.sendAllDeleteBroadcast(mContext1, false)
+                    }
                 })
     }
 
-    private fun delete(taskUrl: String, isDeleteFile: Boolean, isCancelNotify: Boolean) {
-        val downloadTaskBean = DownloadManager.instance.getDownloadTask(taskUrl) ?: return
+    private fun delete(taskUrl: String, isDeleteFile: Boolean) {
+        val downloadTaskBean = DownloadManager.instance.getDownloadTask(taskUrl)
+        if (downloadTaskBean == null) {
+            DownloadTaskFileChangeLister.sendDeleteBroadcast(mContext1, downloadTaskBean, false)
+            return
+        }
         downloadTaskLists.remove(downloadTaskBean)
         TaskManager.instance.delete(downloadTaskBean.url)
         AppDbHelper.instance.deleteSingleMission(downloadTaskBean)
@@ -281,16 +301,51 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                         if (isDeleteFile) {
                             FsUtils.deleteFileOrDir(downloadTaskBean.absolutePath)
                         }
-                        if (isCancelNotify) {
+                        if (downloadTaskBean.showNotification) {
                             notifyHelper.notificationManager.cancel(downloadTaskBean.notificationId)
                         }
-                        DownloadTaskDeleteLister.sendDeleteBroadcast(mContext1, TaskDeleteStatusEvent(TaskDeleteStatusEvent.Status.DELETE_SINGLE, downloadTaskBean))
                         DownloadTaskChangeLister.sendChangeBroadcast(mContext1, downloadTaskBean.apply {
                             this.downloadTaskStatusType = DownloadTaskStatusType.Delete
                         })
+                        DownloadTaskFileChangeLister.sendDeleteBroadcast(mContext1, downloadTaskBean, true)
                     }
 
-                    override fun rxOnError(e: Exception) = Unit
+                    override fun rxOnError(e: Exception) {
+                        DownloadTaskFileChangeLister.sendDeleteBroadcast(mContext1, downloadTaskBean, false)
+                    }
+                })
+    }
+
+    private fun renameTaskFile(taskUrl: String, fileName: String) {
+        val downloadTaskBean = DownloadManager.instance.getDownloadTask(taskUrl)
+        if (downloadTaskBean == null || !FsUtils.exists(downloadTaskBean.absolutePath)
+                || downloadTaskBean.downloadTaskStatusType != DownloadTaskStatusType.Success
+                || fileName.isEmpty()) {
+            DownloadTaskFileChangeLister.sendRenameBroadcast(mContext1, downloadTaskBean, false)
+            return
+        }
+        if (fileName == File(downloadTaskBean.absolutePath).name) {
+            DownloadTaskFileChangeLister.sendRenameBroadcast(mContext1, downloadTaskBean, true)
+            return
+        }
+        val newFile = FsUtils.renameFile(File(downloadTaskBean.absolutePath), fileName)
+        if (!FsUtils.exists(newFile)) {
+            DownloadTaskFileChangeLister.sendRenameBroadcast(mContext1, downloadTaskBean, false)
+            return
+        }
+        downloadTaskBean.absolutePath = newFile!!.absolutePath
+        AppDbHelper.instance
+                .createOrUpdateDownloadTask(downloadTaskBean)
+                .compose(RxObservableTransformer.io_main())
+                .compose(RxObservableTransformer.errorResult())
+                .subscribe(object : RxSubscriber<Long>() {
+                    override fun rxOnNext(t: Long) {
+                        DownloadTaskFileChangeLister.sendRenameBroadcast(mContext1, downloadTaskBean, true)
+                    }
+
+                    override fun rxOnError(e: Exception) {
+                        DownloadTaskFileChangeLister.sendRenameBroadcast(mContext1, downloadTaskBean, false)
+                    }
                 })
     }
 
@@ -334,8 +389,8 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                         .setCategory(NotificationCompat.CATEGORY_PROGRESS)
                         .setShowWhen(false)
         downloadIngNotification?.apply {
-            downloadTaskBean.shortName?.let {
-                this.setContentTitle(it)
+            if (!downloadTaskBean.notificationTitle.isNullOrEmpty()) {
+                this.setContentTitle(downloadTaskBean.notificationTitle)
             }
             this.setContentText(CommonUtils.downloadStateNotificationInfo(mContext1, downloadTaskBean))
             this.setProgress(downloadTaskBean.totalLength.toInt(), downloadTaskBean.currentOffset.toInt(), false)
@@ -351,8 +406,8 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                         .setOngoing(false)
                         .setAutoCancel(true)
         downloadCompatNotification?.apply {
-            downloadTaskBean.shortName?.let {
-                this.setContentTitle(it)
+            if (!downloadTaskBean.notificationTitle.isNullOrEmpty()) {
+                this.setContentTitle(downloadTaskBean.notificationTitle)
             }
             this.setContentText(CommonUtils.downloadStateNotificationInfo(mContext1, downloadTaskBean))
             notifyHelper.notificationManager.cancel(downloadTaskBean.notificationId)
@@ -367,8 +422,8 @@ class DownloadServiceAssistUtils(private val mContext1: Context, clazz: Class<*>
                         .setOngoing(false)
                         .setAutoCancel(true)
         downloadFailedNotification?.apply {
-            downloadTaskBean.shortName?.let {
-                this.setContentTitle(it)
+            if (!downloadTaskBean.notificationTitle.isNullOrEmpty()) {
+                this.setContentTitle(downloadTaskBean.notificationTitle)
             }
             this.setContentText(CommonUtils.downloadStateNotificationInfo(mContext1, downloadTaskBean))
             notifyHelper.notificationManager.cancel(downloadTaskBean.notificationId)
